+++ packages/extension-web-ui/src/Popup/Home/Earning/EarningPositionDetail/desktop/RewardInfoDesktopPart.tsx
@@ -6,13 +6,11 @@ import { NotificationType } from '@subwallet/extension-base/background/KoniTypes
 import { _STAKING_CHAIN_GROUP } from '@subwallet/extension-base/services/earning-service/constants';
 import { EarningRewardHistoryItem, YieldPoolType, YieldPositionInfo } from '@subwallet/extension-base/types';
 import { isSameAddress } from '@subwallet/extension-base/utils';
-import { BaseModal } from '@subwallet/extension-web-ui/components';
 import { EarningRewardsHistoryModal } from '@subwallet/extension-web-ui/components/Modal/Earning/EarningRewardsHistoryModal';
-import { BN_ZERO, CLAIM_REWARD_TRANSACTION, DEFAULT_CLAIM_REWARD_PARAMS, TRANSACTION_YIELD_CLAIM_MODAL } from '@subwallet/extension-web-ui/constants';
+import { BN_ZERO, CLAIM_REWARD_TRANSACTION, DEFAULT_CLAIM_REWARD_PARAMS } from '@subwallet/extension-web-ui/constants';
 import { ScreenContext } from '@subwallet/extension-web-ui/contexts/ScreenContext';
+import { TransactionModalContext } from '@subwallet/extension-web-ui/contexts/TransactionModalContextProvider';
 import { useSelector, useTranslation, useYieldRewardTotal } from '@subwallet/extension-web-ui/hooks';
-import Transaction from '@subwallet/extension-web-ui/Popup/Transaction/Transaction';
-import ClaimReward from '@subwallet/extension-web-ui/Popup/Transaction/variants/ClaimReward';
 import { AlertDialogProps, ThemeProps } from '@subwallet/extension-web-ui/types';
 import { getReformatedAddressRelatedToChain, openInNewTab } from '@subwallet/extension-web-ui/utils';
 import { ActivityIndicator, Button, Icon, ModalContext, Number } from '@subwallet/react-ui';
@@ -41,8 +39,9 @@ function Component ({ className, closeAlert, compound,
   transactionFromValue }: Props) {
   const { t } = useTranslation();
   const navigate = useNavigate();
-  const { activeModal, inactiveModal } = useContext(ModalContext);
+  const { activeModal } = useContext(ModalContext);
   const { isWebUI } = useContext(ScreenContext);
+  const { claimRewardModal } = useContext(TransactionModalContext);
 
   const { currentAccountProxy } = useSelector((state) => state.accountState);
   const chainInfoMap = useSelector((state) => state.chainStore.chainInfoMap);
@@ -92,7 +91,7 @@ function Component ({ className, closeAlert, compound,
 
       // todo: open Modal
       if (isWebUI) {
-        activeModal(TRANSACTION_YIELD_CLAIM_MODAL);
+        claimRewardModal.open();
       } else {
         navigate('/transaction/claim-reward');
       }
@@ -108,10 +107,7 @@ function Component ({ className, closeAlert, compound,
         }
       });
     }
-  }, [type, isDAppStaking, total, setClaimRewardStorage, slug, transactionChainValue, transactionFromValue, isWebUI, activeModal, navigate, openAlert, t, closeAlert]);
-  const handleCloseClaim = useCallback(() => {
-    inactiveModal(TRANSACTION_YIELD_CLAIM_MODAL);
-  }, [inactiveModal]);
+  }, [type, isDAppStaking, total, setClaimRewardStorage, slug, transactionChainValue, transactionFromValue, isWebUI, claimRewardModal, navigate, openAlert, t, closeAlert]);
 
   const onOpenRewardsHistoryModal = useCallback(() => {
     activeModal(rewardsHistoryModalId);
@@ -206,20 +202,6 @@ function Component ({ className, closeAlert, compound,
         onClickViewExplore={onClickViewExplore}
         rewardHistories={rewardHistories}
       />
-      <BaseModal
-        className={'right-side-modal'}
-        destroyOnClose={true}
-        id={TRANSACTION_YIELD_CLAIM_MODAL}
-        onCancel={handleCloseClaim}
-        title={t('Claim rewards')}
-      >
-        <Transaction
-          modalContent={isWebUI}
-          modalId={TRANSACTION_YIELD_CLAIM_MODAL}
-        >
-          <ClaimReward />
-        </Transaction>
-      </BaseModal>
     </>
   );
 }