+++ packages/extension-web-ui/src/components/Field/AddressInputNew.tsx
@@ -401,6 +401,7 @@ function Component (props: Props, ref: ForwardedRef<AddressInputRef>): React.Rea
     <>
       <div className={CN(className, '-input-container')}>
         <AutoComplete
+          disabled={disabled}
           dropdownRender={dropdownRender}
           fieldNames={autoCompleteFieldNames}
           listHeight={dropdownListHeight}
@@ -479,6 +480,7 @@ function Component (props: Props, ref: ForwardedRef<AddressInputRef>): React.Rea
             showAddressBook &&
             (
               <Button
+                disabled={disabled}
                 icon={(
                   <Icon
                     phosphorIcon={Book}